Add files via upload
Cohen_kappa: This notebook applies the Cohen Kappa evaluation metric to calculate the inter - reliability in the decisions between reviewers.
Disagreement in round1: This shows the percentage of reviews for any pair of reviewers that they made different decisions.
Levenshtein_distance: This compares any pair of handles and calculates the number of edits to convert one to the other. 
Total agreement between reviewers: This compares the percentage of reviews where any pair of reviewers made similar judgements.